Move `patch` from `execute` to a separate module
