+++ collector/src/benchmark/patch.rs
@@ -1,0 +1,70 @@
+use crate::command_output;
+use database::PatchName;
+use std::hash;
+use std::path::{Path, PathBuf};
+use std::process::Command;
+
+#[derive(Debug, Clone)]
+pub struct Patch {
+    pub(crate) index: usize,
+    pub name: PatchName,
+    path: PathBuf,
+}
+
+impl PartialEq for Patch {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+    }
+}
+
+impl Eq for Patch {}
+
+impl hash::Hash for Patch {
+    fn hash<H: hash::Hasher>(&self, h: &mut H) {
+        self.name.hash(h);
+    }
+}
+
+impl Patch {
+    pub fn new(path: PathBuf) -> Self {
+        assert!(path.is_file());
+        let (index, name) = {
+            let file_name = path.file_name().unwrap().to_string_lossy();
+            let mut parts = file_name.split("-");
+            let index = parts.next().unwrap().parse().unwrap_or_else(|e| {
+                panic!(
+                    "{:?} should be in the format 000-name.patch, \
+                     but did not start with a number: {:?}",
+                    &path, e
+                );
+            });
+            let mut name = parts.fold(String::new(), |mut acc, part| {
+                acc.push_str(part);
+                acc.push(' ');
+                acc
+            });
+            let len = name.len();
+            // take final space off
+            name.truncate(len - 1);
+            let name = name.replace(".patch", "");
+            (index, name)
+        };
+
+        Patch {
+            path: PathBuf::from(path.file_name().unwrap().to_str().unwrap()),
+            index,
+            name: name.as_str().into(),
+        }
+    }
+
+    pub fn apply(&self, dir: &Path) -> anyhow::Result<()> {
+        log::debug!("applying {} to {:?}", self.name, dir);
+
+        let mut cmd = Command::new("git");
+        cmd.current_dir(dir).args(&["apply"]).arg(&*self.path);
+
+        command_output(&mut cmd)?;
+
+        Ok(())
+    }
+}
